run-blktests: run only one job for one PR
KPD updates the PR when the new version of the patch series is posted.
However, KPD also updates the PR when a new commit is made to the base
branch. This keeps on updating PRs even when new patch version is not
posted. Currently run-blktests workflow triggers a job for every PR
update, then it creates a job for every commit to the base branch (e.g.,
Linux master branch). This results in too many jobs.

To avoid the too many jobs, limit the number of jobs for each PR.
Specify the concurrency keyword to run only one job at a time. When
multiple jobs are created, older, pending jobs are canceled.

Signed-off-by: Shin'ichiro Kawasaki <shinichiro.kawasaki@wdc.com>