run-blktests.yml: mark blktests completion
When blktests CI run failed due to test system reason, reports of such
failures are meaningless for kernel developers. The failure reports
should be sent out only when the blktests run completed.

To decide KPD should send out e-mail reports or not, add a message log
in the GitHub workflow to mark completion of the blktests run. While at
it, modify the message to mark failure test case list start so that the
marks for KPD have the common prefix "KPD:".

Signed-off-by: Shin'ichiro Kawasaki <shinichiro.kawasaki@wdc.com>